+++ HISTORY.md
@@ -15,15 +15,17 @@
 
  * Add clarifying comments in problems/ implementations
    ([#276](https://github.com/mlpack/ensmallen/pull/276)).
-
  * CheckArbitraryFunctionTypeAPI extended for MOO support
    ([#283](https://github.com/mlpack/ensmallen/pull/283)).
 
  * Refactor NSGA2
    ([#263](https://github.com/mlpack/ensmallen/pull/263)).
-
+ 
  * Add Indicators for Multiobjective optimizers
    ([#285](https://github.com/mlpack/ensmallen/pull/285)).
+
+ * Add bounds of search space for problem functions.
+   ([#276](https://github.com/mlpack/ensmallen/pull/279)).
 
 ### ensmallen 2.16.1: "Severely Dented Can Of Polyurethane"
 ###### 2021-03-02